# swagger.yaml
openapi: 3.0.0
info:
  description: |
    This is souvershop server.
  version: "1.0.0"
  title: Swagger souvershop
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://souvershop-ff3c13f9139a.herokuapp.com/api/"
  - url: "http://localhost:8888/api/"
...
paths:
...
  /product:
    post:
      tags:
        - product
      summary: Create a new product
      description: Create a new product
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product_name:
                  type: string
                  format: string
                product_description:
                  type: string
                  format: string
                product_price:
                  type: number
                  format: float
                product_quantity:
                  type: number
                  format: integer
                product_image:
                  type: string
                  format: string
                product_category:
                  type: number
                  format: integer
              required:
                - product_name
                - product_description
                - product_price
                - product_quantity
                - product_image
                - product_category
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/product"
        "400":
          description: Invalid request
    get:
      tags:
        - product
      summary: Get all products
      description: Get all products
      operationId: getProducts
      security:
        - BearerAuth: []
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/product"
        "400":
          description: Invalid request
  /product/{product_id}:
    get:
      tags:
        - product
      summary: Get product's detail by id
      description: Get product by id
      operationId: getProductById
      parameters:
        - name: product_id
          in: path
          description: ID of product to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/product"
        "400":
          description: Invalid request
    put:
      tags:
        - product
      summary: Update product's detail by id
      description: Update product by id
      operationId: updateProductById
      security:
        - BearerAuth: []
      parameters:
        - name: product_id
          in: path
          description: ID of product to return
          required: true
          schema:
            type: number
            format: int2
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product_name:
                  type: string
                  format: string
                product_description:
                  type: string
                  format: string
                product_price:
                  type: number
                  format: float
                product_quantity:
                  type: number
                  format: integer
                product_image:
                  type: string
                  format: string
                product_category:
                  type: number
                  format: integer
              required:
                - product_name
                - product_description
                - product_price
                - product_quantity
                - product_image
                - product_category
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/product"
        "400":
          description: Invalid request
    delete:
      tags:
        - product
      summary: Delete product by id
      description: Delete product by id
      operationId: deleteProductById
      security:
        - BearerAuth: []
      parameters:
        - name: product_id
          in: path
          description: ID of product to return
          required: true
          schema:
            type: number
            format: int2
      responses:
        "200":
          description: successful
...
components:
  schemas:
...
    product:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        product_name:
          type: string
          format: string
        product_description:
          type: string
          format: string
        product_price:
          type: number
          format: float
        product_quantity:
          type: number
          format: integer
        product_image:
          type: string
          format: string
        product_category:
          type: number
          format: integer
        create_at:
          type: string
          format: date-time
        update_at:
          type: string
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT